Split MethodAnalyzer into smaller Classes
